requirements.txt
Faker==37.12.0
apache-airflow-providers-postgres
pandas
matplotlib
seaborn
pyspark==3.5.0